Enable ELF32 and ELF64
Enabled ELF32/64 support in SBL
Tested UEFI payload 32bit and 64bit ELF format.

+++ BootloaderCommonPkg/Include/Library/UniversalPayloadLib.h
@@ -15,8 +15,8 @@
 
 typedef struct {
   UPLD_INFO_HEADER                Info;
-  UINT16                          Machine;
-  UINT16                          ImageCount;
+  UINT32                          Machine;
+  UINT32                          ImageCount;
   UNIVERSAL_PAYLOAD_ENTRYPOINT    EntryPoint;
   PAYLOAD_IMAGE_ENTRY             LoadedImage[MAX_PLD_IMAGE_ENTRY];
 } LOADED_PAYLOAD_INFO;
